Fix -Wall errors

+++ src/data/ZtaFrameBuffer.cpp
@@ -78,7 +78,7 @@ int ZtaFrameBuffer::createBuffer()
         // Process each row
         for (int row = 0; row < frame.height; row++)
         {
-            if (row >= frame.pixelSets.size())
+            if (row >= (int)frame.pixelSets.size())
             {
                 std::cerr << "ERROR: Row " << row << " exceeds ZtaPixelSet count!" << std::endl;
                 continue;
